test: #240 split oversized test files, verb-first helper renames
Review findings 1-2 + naming: advanced-query handler tests move to
server_advanced_query_test.go (server_test.go 537->331); the QueryRequest
JSON cluster moves to types_query_request_test.go (types_test.go 915->676,
remaining overage pre-existing); sortTestCache->buildSortTestCache,
mixedRankOracle->buildMixedRankOracle; gofmt comment-alignment normalization
on touched files.
